refactor(Logo): Extract Logo from Layout
