+++ src/index.js
@@ -9,6 +9,6 @@ export { default as Icon } from './Icon'
 export { default as Input } from './Input'
 export { default as Label } from './Label'
 export { default as Layout } from './Layout'
-export { default as Logo } from './Layout/LayoutTopbar/Logo'
+export { default as Logo } from './Logo'
 export { default as Message } from './Message'
 export { default as NotificationCenter } from './NotificationCenter'
